test/web-platform-tests/to-upstream/FileAPI/blob/Blob-close.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Blob.close</title>
<link rel=help href="http://dev.w3.org/2006/webapi/FileAPI/#dfn-close">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
"use strict";

test(() => {
  const blob = new Blob(["TEST"]);
  assert_false(blob.isClosed, "Blob should not be closed right after construction");
  blob.close();
  assert_true(blob.isClosed, "Blob should be closed after being close()'d");
});
</script>
